+++ webapp/pages/projects.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="projects-container">
     <title-card title="Projects">
-      <form class="project-form" @submit.prevent="submitProject">
+      <form class="project-form" @submit.prevent="updateProject">
         <label :class="`label ${errors.name ? 'error' : ''}`" for="input">Give your project a title</label>
         <input :class="errors.name ? 'error' : ''" type="text" v-model="curProject.name" @blur="onNameBlur" />
         <label class="label" for="input">Describe the project</label>
@@ -21,47 +21,54 @@
       <p>Go to <nuxt-link to="/priorities"><button class="continue">Priorities</button></nuxt-link> when you're done with adding projects</p>
     </title-card>
     <div class="project-list">
-      <card v-for="projectId in Object.keys(projectData).sort((a, b) => projectData[b].createdAt - projectData[a].createdAt)" :key="projectId">
-        <div class="project-row">
-          <div>
-            <h3 class="project-name">
-              {{ projectData[projectId].name }}
-            </h3>
-            <div
-              v-if="categoriesData && categoriesData[projectData[projectId].category]"
-              class="project-category"
-            >
-              <category-pill :color="categoriesData[projectData[projectId].category].color">
-                {{ categoriesData[projectData[projectId].category].name }}
-              </category-pill>
+      <draggable v-model="projectsDisplayedModel" v-bind="dragOptions" @start="drag=true" @end="onDragEnd" handle=".handle">
+        <transition-group type="transition" :name="!drag ? 'flip-list' : null">
+          <card v-for="project in projectsDisplayed" :key="project.id">
+            <div class="project-with-handle">
+              <div class="handle-container"><v-icon name="grip-vertical" width="10" height="16" class="handle" /></div>
+              <div class="project-row">
+                <div>
+                  <h3 class="project-name">
+                    {{ project.name }}
+                  </h3>
+                  <div
+                    v-if="categoriesData && categoriesData[project.category]"
+                    class="project-category"
+                  >
+                    <category-pill :color="categoriesData[project.category].color">
+                      {{ categoriesData[project.category].name }}
+                    </category-pill>
+                  </div>
+                </div>
+                <div class="buttons">
+                  <button
+                    class="secondary icon-only complete"
+                    aria-label="Mark Project as Complete"
+                    @click="completeProject(project.id)"
+                  >
+                    <v-icon name="check"/>
+                  </button>
+                  <button
+                    class="secondary icon-only"
+                    aria-label="Update Project"
+                    @click="updateProject(project.id)"
+                  >
+                    <v-icon name="edit"/>
+                  </button>
+                  <button
+                    class="secondary icon-only delete"
+                    aria-label="Delete Project"
+                    @click="removeProject(project.id)"
+                  >
+                    <v-icon name="trash"/>
+                  </button>
+                </div>
+              </div>
             </div>
-          </div>
-          <div class="buttons">
-            <button
-              class="secondary icon-only complete"
-              aria-label="Mark Project as Complete"
-              @click="completeProject(projectId)"
-            >
-              <v-icon name="check"/>
-            </button>
-            <button
-              class="secondary icon-only"
-              aria-label="Update Project"
-              @click="updateProject(projectId)"
-            >
-              <v-icon name="edit"/>
-            </button>
-            <button
-              class="secondary icon-only delete"
-              aria-label="Delete Project"
-              @click="removeProject(projectId)"
-            >
-              <v-icon name="trash"/>
-            </button>
-          </div>
-        </div>
-        <div v-if="projectData[projectId].description" class="project-description" v-html="linkifyHtml(projectData[projectId].description)" />
-      </card>
+            <div v-if="project.description" class="project-description" v-html="linkifyHtml(project.description)" />
+          </card>
+        </transition-group>
+      </draggable>
     </div>
   </div>
 </template>
@@ -69,13 +76,14 @@
 <script>
 import { mapState } from 'vuex'
 import linkifyHtml from 'linkify-html';
+import Draggable from 'vuedraggable'
 import TitleCard from '../components/TitleCard.vue';
 import Card from '../components/Card.vue'
 import CategoryPill from '../components/CategoryPill.vue'
 
 export default {
   name: 'Projects',
-  components: { TitleCard, Card, CategoryPill },
+  components: { TitleCard, Card, CategoryPill, Draggable },
   data () {
     return {
       curProject: {
@@ -88,6 +96,7 @@ export default {
         name: false,
         category: false,
       },
+      drag: false,
     }
   },
   computed: {
@@ -97,6 +106,28 @@ export default {
     }),
     cannotSubmit() {
       return !this.curProject.name || this.curProject.name.trim === '' || !this.curProject.category || this.curProject.category.trim === ''
+    },
+    dragOptions() {
+      return {
+        animation: 200,
+        group: 'Projects',
+        ghostClass: "ghost"
+      };
+    },
+    projectsDisplayedModel: {
+      get() {
+        this.projectsDisplayed = this.projectData ? Object.keys(this.projectData).map(id => this.projectData[id]) : [];
+        return this.projectsDisplayed.sort((a, b) => (b.order || b.createdAt) - (a.order || a.createdAt));
+      },
+      set(value) {
+        this.projectsDisplayed = value;
+      }
+    }
+  },
+  watch: { 
+    projectData: function(newProjectData) {
+      this.projectsDisplayed = newProjectData ? Object.keys(newProjectData).map(id => newProjectData[id]) : [];
+      this.projectsDisplayed = this.projectsDisplayed.sort((a, b) => (b.order || b.createdAt) - (a.order || a.createdAt));
     }
   },
   mounted () {
@@ -113,9 +144,9 @@ export default {
     onCategoryBlur() {
       this.errors.category = !this.curProject.category || this.curProject.category.trim() === '';
     },
-    submitProject () {
+    updateProject () {
       if (this.cannotSubmit) return;
-      this.$store.dispatch('projects/submitProject', this.curProject);
+      this.$store.dispatch('projects/updateProject', this.curProject);
       this.formState = 'Add'
       this.curProject = {
         name: '',
@@ -124,7 +155,7 @@ export default {
       }
     },
     completeProject (id) {
-      this.$store.dispatch('projects/submitProject', { id, complete: true });
+      this.$store.dispatch('projects/updateProject', { id, complete: true });
       this.formState = 'Add'
       this.curProject = {
         name: '',
@@ -138,7 +169,25 @@ export default {
       window.scrollTo(0, 0);
     },
     removeProject (id) {
-      this.$store.dispatch('projects/removeProject', id)
+      if (confirm('Are you sure you want to delete this project?')) {
+        this.$store.dispatch('projects/removeProject', id);
+      }
+    },
+    onDragEnd (e) {
+      const projects = Object.keys(this.projectData).map(id => this.projectData[id]).sort((a, b) => (b.order || b.createdAt) - (a.order || a.createdAt));
+      this.drag = false
+      const curIndex = e.oldIndex;
+      const newIndex = e.newIndex;
+      const curProject = projects[curIndex];
+      let newOrder;
+      const existingProjectOrder = projects[newIndex].order || projects[newIndex].createdAt;
+      if (newIndex > curIndex) {
+        newOrder = existingProjectOrder - 1;
+      } else {
+        newOrder = existingProjectOrder + 1;
+      }
+
+      this.$store.dispatch('projects/updateProject', { ...curProject, order: newOrder });
     },
   }
 }
@@ -168,6 +217,20 @@ p {
 
 .card {
   margin-bottom: 25px;
+}
+
+.project-with-handle {
+  display: flex;
+  align-items: center;
+}
+
+.handle-container {
+  margin-right: 20px;
+  cursor: pointer;
+}
+
+.project-description {
+  margin: 10px 0 0 30px;
 }
 
 .project-row {
